requirements.txt
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
chromedriver-autoinstaller==0.6.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
fastapi==0.119.1
h11==0.16.0
httptools==0.7.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
loguru==0.7.3
markdown-it-py==4.0.0
mdurl==0.1.2
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PySocks==1.7.1
pytest==8.4.2
pytest-mock==3.15.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
selenium==4.37.0
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.48.0
tenacity==9.1.2
trio==0.31.0
trio-websocket==0.12.2
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
win32_setctime==1.2.0
wsproto==1.2.0

## Excel / preview support
pandas>=2.0
openpyxl>=3.1
xlrd>=2.0

# Google Sheets integration
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0

# PDF generation
reportlab>=4.0.0

rapidfuzz>=2.0
